First raising passes: pass infrastructure, typed constants, property sugar
The pass list arrives (IIrPass + IrPasses.Default, CheckInvariant after
every pass in debug builds — the ordered list IS the architecture
document), with the first two passes targeting the parity scoreboard's
top buckets:

TypedConstantsPass retypes integer constants by consuming position
(return type, store targets, call arguments, comparisons) — IL has no
bool or char constants, so a 0 returned from a bool method IS false;
recovering it is raising, the inverse of constant lowering.

PropertySugarPass raises get_/set_ accessor calls to typed
LoadProperty/StoreProperty nodes (indexers included) — tree rewrites
via the new IrNode.DetachChildren primitive, not name surgery at print
time. The printer renders them with a shared ReceiverText helper that
also fixes value-type receivers arriving by address ((ref x).M() now
prints x.M()).

The typed-constants pass immediately caught an import bug: ldc.i4.N
constants boxed a ushort (enum subtraction yields the enum's
underlying type) while claiming Int32 — the pass's 'is int' check
refused to touch values that lied about their own type. Fixed at the
importer.

Parity: 30.72% -> 37.07% (+6.35 points; predicted buckets closed:
bool-as-int 505, property sugar 217+, address-receivers 69+).

+++ src/DotnetInspector.Decompiler/Pipeline/Ir/IrImporter.cs
@@ -373,7 +373,10 @@ public static class IrImporter
                     break;
 
                 case >= ILOpCode.Ldc_i4_m1 and <= ILOpCode.Ldc_i4_8:
-                    stack.Push(new Constant(opcode - ILOpCode.Ldc_i4_0, TypeRef.CoreLib("System", "Int32")));
+                    // Enum subtraction yields the enum's underlying ushort —
+                    // box an actual int or every typed-position check downstream
+                    // sees a value that lies about its own type.
+                    stack.Push(new Constant((int)(opcode - ILOpCode.Ldc_i4_0), TypeRef.CoreLib("System", "Int32")));
                     break;
                 case ILOpCode.Ldc_i4_s:
                     stack.Push(new Constant((int)(sbyte)reader.ReadILByte(), TypeRef.CoreLib("System", "Int32")));
